requirements.txt
ffmpeg-python
numpy
python-dotenv
sarvamai
faiss-cpu
sentence-transformers
transformers
google-generativeai
